irlobby_backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
celery==5.4.0
channels==4.3.1
channels-redis==4.2.0
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
dj-database-url==3.0.1
Django>=4.2,<5      # compatible with local Python 3.9 LTS
django-cors-headers==4.4.0
django-waffle==5.0.0
django-environ==0.11.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-allauth==65.4.1
factory_boy==3.3.0
requests==2.32.3
requests-oauthlib==2.0.0
oauthlib==3.2.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
packaging==25.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
PyJWT==2.10.1
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-django==4.8.0
redis==5.0.7
ruff==0.6.8
sentry-sdk==2.8.0
black==24.8.0
mypy==1.11.1
pre-commit==3.8.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing-extensions==4.15.0
tzdata==2025.2
whitenoise==6.9.0
Pillow>=10.0.0,<11.0.0
qrcode>=7.5,<8.0
stripe>=12.0.0,<13.0.0
zope.interface==7.2
bandit==1.7.9
safety==3.2.4
